adma_tools_cpp: code cleanup and add checks/linter
- also bump all packages to Cmake v3.8, C++17, version 2.4 and add consistent compiler options

+++ adma_core_lib/include/adma_core_lib/parser/parser_utils.hpp
@@ -57,7 +57,7 @@ inline T rad2Deg(T rad)
 bool isLittleEndian();
 
 template<typename T_source>
-T_source extractValue(const std::array<char, 856> & buffer, size_t offset, int lengthBytes)
+T_source extractValue(const std::array<char, 856> & buffer, size_t offset, size_t lengthBytes)
 {
   if (lengthBytes > sizeof(T_source)) {
     RCLCPP_ERROR(rclcpp::get_logger("adma_core_lib::parser_utils"), "Invalid datatype");
